# qim-server/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.AIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  handler.DescribeImageRequest:
    properties:
      image_url:
        type: string
      instruction:
        description: 可选：自定义识别指令，默认"识别图片内容并详细描述"
        type: string
    required:
    - image_url
    type: object
...
  handler.TranslateImageRequest:
    properties:
      image_url:
        type: string
      target_lang:
        description: 默认 zh
        type: string
    required:
    - image_url
    type: object
  handler.TranslateTextRequest:
    properties:
      source_lang:
        description: auto/zh/en/...
        type: string
      target_lang:
        description: zh/en/ja/ko/fr/de
        type: string
      text:
        type: string
    required:
    - target_lang
    - text
    type: object
...
host: localhost:8080
info:
  contact:
    name: QIM Team
  description: QIM 智能办公平台后端 API 文档
  title: QIM Server API
  version: "2.0"
paths:
...
  /api/v1/ai/describe-image:
    post:
      consumes:
      - application/json
      description: 视觉模型识别/描述图片内容，支持自定义识别指令
      parameters:
      - description: 图片识别请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.DescribeImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/handler.AIResponse'
        "400":
          description: 参数错误或未配置视觉路由
          schema:
            $ref: '#/definitions/handler.AIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.AIResponse'
      summary: 图片识别/描述
      tags:
      - AI
...
  /api/v1/ai/translate:
    post:
      consumes:
      - application/json
      description: 将文本翻译为目标语言
      parameters:
      - description: 翻译请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.TranslateTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/handler.AIResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.AIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.AIResponse'
      summary: 翻译文本
      tags:
      - AI
  /api/v1/ai/translate/image:
    post:
      consumes:
      - application/json
      description: 上传图片地址，视觉模型识别图中文字并翻译为目标语言
      parameters:
      - description: 图片翻译请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.TranslateImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/handler.AIResponse'
        "400":
          description: 参数错误或图片无可翻译文字
          schema:
            $ref: '#/definitions/handler.AIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.AIResponse'
      summary: 图片翻译
      tags:
      - AI
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"